Выносит временные данные в отдельный файл
