+++ js/data.js
@@ -1,0 +1,34 @@
+const getDataImagesArray = () => {
+  const COMMENTS = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  ];
+
+  const COMMENTATORS = [
+    'Иван',
+    'Сергей',
+    'Ирина',
+    'Владимир',
+    'Виктория',
+    'Николай',
+    'Альберт',
+    'Снежана',
+    'Захар'
+  ];
+
+  const DESCRIPTION = [
+    'Ничего не обычного. Просто фото',
+    'Фото хорошее, но чего-то не хватает',
+    'Однозначно лайк и репост',
+    'Что тут вообще изображено?',
+    'На любителя, если честно, я бы так не сделал'
+  ];
+
+  return {COMMENTS, COMMENTATORS, DESCRIPTION};
+};
+
+export { getDataImagesArray };
